backend/requirements.txt
flask==3.1.2
flask-cors==6.0.0
httpx==0.28.1
anthropic==0.49.0
PyJWT==2.12.1
bcrypt==4.3.0
python-dotenv==1.1.0
gunicorn==23.0.0
chromadb>=1.5.0
sentence-transformers>=3.4.0
PyMuPDF>=1.25.0
pdfplumber>=0.11.0
resend>=2.0.0
